Add strict flag to .parse(), enabled by default
This adds an optional second param named options to the .parse() method.
Currently the only option is `strict` which defaults to `true` unless explicitly set to `false`

Setting strict mode to `false` prevents it from throwing errors on invalid input, and it instead it simply returns a "best guess".

This also adds several additional checks when strict mode is enabled:
* no more than one image candidate for any given descriptor
* no more than one descriptor per image candidate
* height descriptor gets an explicit error message (or is silently accepted when strict mode is disabled)

Because of the duplicate check, the deepUnique function is no longer needed when strict mode is enabled, and IMO undesired when strict mode is disabled. Because of this, I removed it.

I also tweaked the existing tests to remove invalid input and added a set of new tests with intentionally invalid input to ensure it throw when strict mode is enabled and does not throw when disabled.