Add strict flag to .stringify(), enabled by default
This adds an optional second param named options to the .stringify() method.
Currently the only option is `strict` which defaults to `true` unless explicitly set to `false`

Setting strict mode to `false` prevents it from throwing errors on invalid input, and it instead it simply returns a "best guess".

I refactored the existing checks into a few standalone methods so that the same checks can be run for both .parse() and .stringify(). In some cases, the error messages were changed slightly.

I removed the set from the .stringify() method, allowing for duplicate image candidates when strict mode is disabled (it is prevented by the check when enabled).

I also tweaked the existing tests to remove invalid input and added a set of new tests with intentionally invalid input to ensure it throw when strict mode is enabled and does not throw when disabled.